requirements.txt
asgiref==3.6.0
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
pycparser==2.21
PyMySQL==1.0.3
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
typing==3.7.4.3
tzdata==2023.3
urllib3==1.26.15
wincertstore==0.2
yarg==0.1.9
channels~=4.0.0